Jebitech_chatbot/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
bcrypt
schedule
python-dotenv
google-generativeai
pandas
cryptography
